requirements.txt
datetime
json
matplotlib
numpy
pandas
requests
seaborn
sklearn
tqdm
python-dotenv